requirements.txt
appnope==0.1.2
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
decorator==5.1.1
flake8==4.0.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
ipython==7.31.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.29
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
WTForms==3.0.1
